+++ exporter/opentelemetry-exporter-prometheus/src/opentelemetry/exporter/prometheus/__init__.py
@@ -66,9 +66,11 @@ from collections import deque
 from itertools import chain
 from json import dumps
 from logging import getLogger
+from os import environ
 from re import IGNORECASE, UNICODE, compile
 from typing import Dict, Sequence, Tuple, Union
 
+from prometheus_client import start_http_server
 from prometheus_client.core import (
     REGISTRY,
     CounterMetricFamily,
@@ -78,6 +80,10 @@ from prometheus_client.core import (
 )
 from prometheus_client.core import Metric as PrometheusMetric
 
+from opentelemetry.sdk.environment_variables import (
+    OTEL_EXPORTER_PROMETHEUS_HOST,
+    OTEL_EXPORTER_PROMETHEUS_PORT,
+)
 from opentelemetry.sdk.metrics import Counter
 from opentelemetry.sdk.metrics import Histogram as HistogramInstrument
 from opentelemetry.sdk.metrics import (
@@ -375,3 +381,21 @@ class _CustomCollector:
         info = InfoMetricFamily(name, description, labels=attributes)
         info.add_metric(labels=list(attributes.keys()), value=attributes)
         return info
+
+
+class _AutoPrometheusMetricReader(PrometheusMetricReader):
+    """Thin wrapper around PrometheusMetricReader used for the opentelemetry_metrics_exporter entry point.
+
+    This allows users to use the prometheus exporter with opentelemetry-instrument. It handles
+    starting the Prometheus http server on the the correct port and host.
+    """
+
+    def __init__(self) -> None:
+        super().__init__()
+
+        # Default values are specified in
+        # https://github.com/open-telemetry/opentelemetry-specification/blob/v1.24.0/specification/configuration/sdk-environment-variables.md#prometheus-exporter
+        start_http_server(
+            port=int(environ.get(OTEL_EXPORTER_PROMETHEUS_PORT, "9464")),
+            addr=environ.get(OTEL_EXPORTER_PROMETHEUS_HOST, "localhost"),
+        )